requirements.txt
Flask==1.1.1
Jinja2==2.10.3
numpy==1.18.1
requests==2.22.0
urllib3==1.25.7
Werkzeug==0.16.0
gunicorn==19.9.0
annoy==1.16.3
Flask==1.1.1
flask-paginate==0.5.5
Flask-SQLAlchemy==2.4.1
flask-webpackext==1.0.0
Jinja2==2.10.3
numpy==1.18.1
Pillow==7.0.0
requests==2.22.0
scikit-learn==0.22.1
sklearn==0.0
https://dl.google.com/coral/python/tflite_runtime-1.14.0-cp36-cp36m-linux_x86_64.whl
Werkzeug==0.16.0
